Reorganize asset plugin directory layout :wrench:
