+++ client/plugins/asset/link.js
