+++ client/plugins/asset/upload.js
